Implemented debating results code (early days...)
I've implemented DebatingResult but not run it or tested it at all.
I've integrated debating into festival_info.yaml and the corresponding
code (sections, points), and I've modified enough code and tests to
get the old tests working.

A report can be generated for 2012 with the code in this commit.

Now to test the debating results code.