Report successfully generated with Debating data
Far from complete.  The total_points in the summary is a hardcoded result just
to get a report printed.  There is no "Debating" section.

However, the points are being added up correctly for each school, including
debating, and the results are being printed nicely.  Here is an example:

  PLC Sydney                 (J: 37  S: 53  T: 90)
    Junior
      Debating                7 points (p, r1, r2a)
      Reading                12 points (2nd place)
      Poetry                  3 points (p)
      Public Speaking        15 points (1st place)
    Senior
      Debating               15 points (p, r1, r2a, qf, sf)
      Reading                 6 points (5th place)
      Poetry                  3 points (p)
      Public Speaking        10 points (3rd place)
      Drama                  12 points (2nd place)
      Current Affairs         3 points (p)
      Religious and ...      12 points (2nd place)

As the code stands now, I could print a report at the end of the festival.
However, it wouldn't contain a detailed account of Debating.

Also, I couldn't create reports _during_ the festival, because I'm pretty sure
that incomplete debating data will cause an error.